simplify server structure
